+++ src-ssr/server.js
@@ -1,95 +1,69 @@
 /**
- * More info about this file:
  * https://v2.quasar.dev/quasar-cli-vite/developing-ssr/ssr-webserver
  *
- * Runs in Node context.
+ * Runs in Node.js context.
  */
 
-/**
- * Make sure to yarn add / npm install (in your project root)
- * anything you import here (except for express and compression).
- */
 import express from 'express'
 import compression from 'compression'
 import {
-  ssrClose,
-  ssrCreate,
-  ssrListen,
-  ssrRenderPreloadTag,
-  ssrServeStaticContent
-} from 'quasar/wrappers'
+  defineSsrClose,
+  defineSsrCreate,
+  defineSsrInjectDevMiddleware,
+  defineSsrListen,
+  defineSsrRenderPreloadTag,
+  defineSsrServeStaticContent
+} from '#q-app'
 
-/**
- * Create your webserver and return its instance.
- * If needed, prepare your webserver to receive
- * connect-like middlewares.
- *
- * Should NOT be async!
- */
-export const create = ssrCreate((/* { ... } */) => {
+export const create = defineSsrCreate(async () => {
   const app = express()
 
-  // attackers can use this header to detect apps running Express
-  // and then launch specifically-targeted attacks
   app.disable('x-powered-by')
 
-  // place here any middlewares that
-  // absolutely need to run before anything else
-  if (process.env.PROD) {
+  if (import.meta.env.QUASAR_PROD) {
     app.use(compression())
   }
 
   return app
 })
 
-/**
- * You need to make the server listen to the indicated port
- * and return the listening instance or whatever you need to
- * close the server with.
- *
- * The "listenResult" param for the "close()" definition below
- * is what you return here.
- *
- * For production, you can instead export your
- * handler for serverless use or whatever else fits your needs.
- */
-export const listen = ssrListen(async ({ app, port, isReady }) => {
-  await isReady()
-  return app.listen(port, () => {
-    if (process.env.PROD) {
-      console.log('Server listening at port ' + port)
+export const injectDevMiddleware = defineSsrInjectDevMiddleware(
+  ({ app }) =>
+    middleware => {
+      app.use(middleware)
+    }
+)
+
+export const listen = defineSsrListen(async ({ app, devHttpsOptions, port }) => {
+  if (import.meta.env.QUASAR_DEV && devHttpsOptions) {
+    const https = await import('node:https')
+    const server = https.createServer(devHttpsOptions, app)
+    return server.listen(port)
+  }
+
+  const http = await import('node:http')
+  const server = http.createServer(app)
+  return server.listen(port, () => {
+    if (import.meta.env.QUASAR_PROD) {
+      console.log(`Server listening at port ${port}`)
     }
   })
 })
 
-/**
- * Should close the server and free up any resources.
- * Will be used on development only when the server needs
- * to be rebooted.
- *
- * Should you need the result of the "listen()" call above,
- * you can use the "listenResult" param.
- *
- * Can be async.
- */
-export const close = ssrClose(({ listenResult }) => {
-  return listenResult.close()
-})
+export const close = defineSsrClose(({ listenResult }) => listenResult.close())
 
-const maxAge = process.env.DEV
-  ? 0
-  : 1000 * 60 * 60 * 24 * 30
+const maxAge = import.meta.env.QUASAR_DEV ? 0 : 1000 * 60 * 60 * 24 * 30
 
-/**
- * Should return middleware that serves the indicated path
- * with static content.
- */
-export const serveStaticContent = ssrServeStaticContent((path, opts) => {
-  return express.static(path, {
-    maxAge,
-    ...opts
-  })
-})
+export const serveStaticContent = defineSsrServeStaticContent(
+  ({ app, resolve }) =>
+    ({ urlPath, pathToServe, opts = {} }) => {
+      const serveFn = express.static(resolve.public(pathToServe), {
+        maxAge,
+        ...opts
+      })
+      app.use(resolve.urlPath(urlPath), serveFn)
+    }
+)
 
 const jsRE = /\.js$/
 const cssRE = /\.css$/
@@ -99,37 +73,33 @@ const gifRE = /\.gif$/
 const jpgRE = /\.jpe?g$/
 const pngRE = /\.png$/
 
-/**
- * Should return a String with HTML output
- * (if any) for preloading indicated file
- */
-export const renderPreloadTag = ssrRenderPreloadTag((file) => {
-  if (jsRE.test(file) === true) {
+export const renderPreloadTag = defineSsrRenderPreloadTag((file) => {
+  if (jsRE.test(file)) {
     return `<link rel="modulepreload" href="${file}" crossorigin>`
   }
 
-  if (cssRE.test(file) === true) {
-    return `<link rel="stylesheet" href="${file}">`
+  if (cssRE.test(file)) {
+    return `<link rel="stylesheet" href="${file}" crossorigin>`
   }
 
-  if (woffRE.test(file) === true) {
+  if (woffRE.test(file)) {
     return `<link rel="preload" href="${file}" as="font" type="font/woff" crossorigin>`
   }
 
-  if (woff2RE.test(file) === true) {
+  if (woff2RE.test(file)) {
     return `<link rel="preload" href="${file}" as="font" type="font/woff2" crossorigin>`
   }
 
-  if (gifRE.test(file) === true) {
-    return `<link rel="preload" href="${file}" as="image" type="image/gif">`
+  if (gifRE.test(file)) {
+    return `<link rel="preload" href="${file}" as="image" type="image/gif" crossorigin>`
   }
 
-  if (jpgRE.test(file) === true) {
-    return `<link rel="preload" href="${file}" as="image" type="image/jpeg">`
+  if (jpgRE.test(file)) {
+    return `<link rel="preload" href="${file}" as="image" type="image/jpeg" crossorigin>`
   }
 
-  if (pngRE.test(file) === true) {
-    return `<link rel="preload" href="${file}" as="image" type="image/png">`
+  if (pngRE.test(file)) {
+    return `<link rel="preload" href="${file}" as="image" type="image/png" crossorigin>`
   }
 
   return ''