+++ src/layouts/MainLayout.vue
@@ -1,19 +1,20 @@
 <template>
-  <q-layout view="lHh Lpr lFf">
-    <q-header elevated>
+  <q-layout view="lHh Lpr lFf" class="bg-dark">
+    <q-header>
       <q-toolbar>
 
         <q-toolbar-title v-if="!$q.platform.is.mobile" class="custom-font">
           Text Encrypter
         </q-toolbar-title>
 
-        <div>v2.0.0
+        <div>v2.1.0
           <a href="https://mobycrypt.com" target="_blank" style="color: white"> </a>
-          | made by <a href="https://mobycrypt.com" target="_blank" style="color: white">mobycrypt.com</a></div>
+          | made by <a href="https://mobycrypt.com" target="_blank" style="color: white">mobycrypt.com</a>
+        </div>
       </q-toolbar>
     </q-header>
 
-    <q-page-container>
+    <q-page-container class="bg-dark">
       <router-view />
     </q-page-container>
   </q-layout>
@@ -21,7 +22,7 @@
 
 <script>
 import { defineComponent, ref } from 'vue'
-import EssentialLink from 'components/EssentialLink.vue'
+import EssentialLink from '@/components/EssentialLink.vue'
 
 export default defineComponent({
   name: 'MainLayout',
@@ -30,12 +31,12 @@ export default defineComponent({
     EssentialLink
   },
 
-  setup () {
+  setup() {
     const leftDrawerOpen = ref(false)
 
     return {
       leftDrawerOpen,
-      toggleLeftDrawer () {
+      toggleLeftDrawer() {
         leftDrawerOpen.value = !leftDrawerOpen.value
       }
     }